Add RotManager to the sprockets-host crate.
An RotManager serializes requests from sessions running in async codebases via
the use of tokio channels. This allows interaction with the host
serialport/Uarts to occur in their own blocking thread, while allowing the rest
of the session code to proceed in tokio tasks.

The RotTransport API is based on the `Uart` type in the sprockets-host crate.
Currently, we test it with a dummy transport that allows injecting errors. The
most important error case is to ensure that if messages take a long time to
process by the RoT, stale messages get ignored. This may lead to a timeout on
the next message, which we force with a zero timeout for the request.

There are a number of unused warnings here. That is expected, as this code is
not yet used yet. The async session networking code will use this in a follow up
commit. Logging will also be added in this or a follow up.