Session: Decouple writing plaintext from flushing ciphertext
This commit builds upon #14, but restructures the `EncryptingBufWriter` so that
operations on plaintext and ciphertext are more clearly delineated. Plaintext is
written into the buffer, and then it is encrypted and flushed when either the
plaintext buffer is full or the caller explicitly requests a flush. Once the
ciphertext is being flushed, data cannot be written to the buffer, until it is
completely flushed.

The behavior described above is identical to the implementation in #14, but the
implementation is more explicit. Rather than relying on a `flush` range to
determine if we have ciphertext, we put the actual buffer into its own 2-state
typestate. The buffer is writable when it is plaintext. Once encrypted, it
becomes an immutable ciphertext frame that includes a 4-byte size header and
trailing authentication tag. Calls to `poll_write` or `poll_flush` will flush
any ciphertext until complete. When the ciphertext has been completely flushed,
the typestate will transition back to an empty plaintext buffer. This ensures at
compile time that the proper operations are taken and makes the code somewhat
easier to read.

As part of this change, `EncryptingBufWriter` now contains an enum,
`AeadWriteBuf`, that contains the 2 typestates of the underlying buffer. To
maintain the compile time typestate guarantees, `self` is taken by
value when `AeadPlaintextBuf::encrypt` is called, so that the user will not
accidentally append plaintext to ciphertext. Unfortunately, due to rust's
semantics, this requires storing `AeadWriteBuf` inside an `Option` so that we
can extract it from `EncryptingBufWriter` and replace it with the new typestate
value. Because of this, we must ensure that we always reset the value of
`self.buf` in `EncryptingBufWriter` to `Some` before we return from an API call.
This prohibits early return in some cases and forces match statements. In short,
we trade off some ergonomics for compile time safety in terms of the typestate
pattern and runtime asserts if we make a major error and don't reset the value
of `self.buf`. We hope this also adds clarity to the code due to its explicitness.

As part of this commit, two other subtle changes were made:

1. When creating a buffer, the user gives the data capacity and tag length. They
no longer explicitly calculate the size of the buffer and include the frame size
header and tag length in their capacity. The intention is to treat the buffer
when in plaintext mode as something where a user writes up their data capacity,
and only treat the header and tag as valid once the ciphertext is
created/sealed.
2. The encryption function no longer returns an `io::Error`. It returns `()` to
clearly indicate that encryption errors should be opaque to prevent side channel
attacks. While the error returned from our current encryption library is opaque,
this prevents subtle errors in the case we change libraries in the future that
don't provide this guarantee.