Session: Decouple reading/decrypting ciphertext from copying plaintext (#19)
* Session: Decouple reading/decrypting ciphertext from copying plaintext

This commit is the companion to #17. It attempts to clarify the stages of
buffered decryption for easier reading. The buffer manipulation itself is now
performed in two typestates in `aead_read_buf`. Reading from the underlying data
source and copying from the buffered reader to the users buffer is performed in
`decrypting_buf_reader`.

This code was made more linear and even clearer with code from @jgallagher. It was a great joint effort!

Co-authored-by: John Gallagher <john@oxidecomputer.com>