Relax measurement requirements
Currently we require measurement appraisal against a valid corpus
set to be successful in order to connect via sprockets. This is
what we want in the ideal world. Unfortunately, we don't live in
the ideal world and we're still working out how the corpus should
be updated. Allow callers to determine how to handle measurement
failures should be handled. This still requires a full verifcation
of the certificate chain which should always work on hardware.