Add cert generation code for testing trust quorum
This PR allows certs to be generated for multiple nodes so that we can
spin up arbitrary test clusters for trust quorum testing in Omicron.

A small change was made to the example `config.kdl` to better enable
code based generation as used in the test.

Depends upon https://github.com/oxidecomputer/pki-playground/pull/126