Allow server accept calls to proceed concurrently (#88)
* Allow server accept calls to proceed concurrently

The sprockets handshake includes both TLS handshake and bidirectional
remote attestation. On real Oxide hardware this results in talking over
IPCC to the SP and then over SPI to the RoT. This introduces delays in
the protocol. And while each message that must reach the RoT may have
to go over IPCC serially (pipelined?), there are other operations that
occur interleaved with these operations, such as the messages sent
between nodes over the bootstrap network. We therefore want to allow the
accept operations as a whole to proceed concurrently, which will speed
up cold boot and also match the behavior we allow on the client side.

As in most async networking code, we wait for each individual connection
by polling the TCP socket acceptor. Once that happens we return a future
that can be polled concurrently or in parallel as desired by the caller.
This matches the tokio-rustls pattern.

A test was added to ensure that the pattern works as we expect with
`spawn`, which we plan to use in Omicron.

Some minor test deduplicaton was also done.